+++ packages/cms/src/containers/Layout.js
@@ -6,6 +6,9 @@ import Content from "./Content";
 import Footer from "./Footer";
 import { SidebarProvider } from "context/sidebar";
 
+// initialize API
+import "api";
+
 const Layout = () => {
   return (
     <div className="c-app c-default-layout">
